+++ crates/full-node/sov-ethereum/src/handlers/mod.rs
@@ -1,7 +1,4 @@
 #![allow(dead_code)]
-
-mod subscribe;
-
 use alloy_primitives::Address;
 use alloy_primitives::{Bytes, B256};
 use alloy_rpc_types::pubsub::SubscriptionKind;
@@ -22,7 +19,6 @@ use sov_modules_api::{RawTx, Spec};
 use sov_sequencer::Sequencer;
 use std::sync::Arc;
 use std::time::Duration;
-pub use subscribe::eth_subscribe;
 
 use crate::to_jsonrpsee_error_object;
 use crate::Ethereum;
@@ -165,4 +161,102 @@ where
     })?;
 
     Ok(tx_hash)
+}
+
+use alloy_rpc_types_eth::pubsub::Params;
+
+#[derive(Debug, Clone, serde::Deserialize)]
+struct EthSubscribe {
+    kind: SubscriptionKind,
+    params: Params,
+}
+
+pub async fn eth_subscribe<S, Seq>(
+    parameters: JRpcParams<'static>,
+    pending: PendingSubscriptionSink,
+    ethereum: Arc<Ethereum<S, Seq>>,
+    _: Extensions,
+) -> jsonrpsee::core::SubscriptionResult
+where
+    S: Spec,
+    Seq: Sequencer<Spec = S>,
+    S::Address: FromVmAddress<EthereumAddress>,
+    Seq::Rt: HasKernel<S> + EthereumAuthenticator<S> + Default + Send + Sync + 'static,
+{
+    let eth_subscribe = match parameters.parse::<EthSubscribe>() {
+        Ok(eth_subscribe) => eth_subscribe,
+        Err(e) => {
+            pending.reject(ErrorObjectOwned::from(e)).await;
+            return Ok(());
+        }
+    };
+
+    let accepted_sink = pending.accept().await?;
+
+    let log_filter = match &eth_subscribe.kind {
+        SubscriptionKind::Logs => match eth_subscribe.params {
+            Params::Logs(filter) => filter,
+            Params::Bool(_) => {
+                return Err(to_jsonrpsee_error_object(
+                    "Boolean parameters are not supported in LOG subscriptions.",
+                    ETH_RPC_ERROR,
+                )
+                .into());
+            }
+            _ => Default::default(),
+        },
+        _ => {
+            return Err(to_jsonrpsee_error_object(
+                "Only LOG subscriptions are supported.",
+                ETH_RPC_ERROR,
+            )
+            .into())
+        }
+    };
+
+    let _task = tokio::spawn(async move {
+        stream_logs(accepted_sink, log_filter, ethereum.clone()).await;
+    });
+
+    Ok(())
+}
+
+async fn stream_logs<S, Seq>(
+    accepted_sink: SubscriptionSink,
+    log_filter: Box<Filter>,
+    _ethereum: Arc<Ethereum<S, Seq>>,
+) where
+    S: Spec,
+    Seq: Sequencer<Spec = S>,
+    S::Address: FromVmAddress<EthereumAddress>,
+    Seq::Rt: HasKernel<S> + EthereumAuthenticator<S> + Default + Send + Sync + 'static,
+{
+    let mut x = 0;
+
+    tokio::time::sleep(Duration::from_millis(1)).await;
+
+    let log = Log {
+        inner: alloy_primitives::Log {
+            address: Address::parse_checksummed("0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045", None)
+                .unwrap(),
+
+            data: LogData::new_unchecked(vec![], Default::default()),
+        },
+        block_hash: None,
+        block_number: None,
+        block_timestamp: None,
+        transaction_hash: None,
+        transaction_index: None,
+        log_index: None,
+        removed: false,
+    };
+
+    let msg = SubscriptionMessage::new(
+        accepted_sink.method_name(),
+        accepted_sink.subscription_id(),
+        &log,
+    )
+    .unwrap();
+
+    accepted_sink.send(msg).await.unwrap();
 }
